Add general_functions sub-module
